requirements.txt
# Core Framework
langchain
langchain-openai
langgraph
langsmith

# Data Processing
pandas
numpy
pydantic
pydantic-settings
python-dotenv

# Database & Cache
redis


# Optimization & Routing
ortools
geopy
osmnx
networkx

# Machine Learning
scikit-learn
xgboost
mlflow

# Document Processing
pypdf2
pytesseract
pillow
holidays
python-multipart
pymupdf

# API & Web
fastapi
uvicorn
httpx
requests

# Dashboard
plotly


# Utilities
tqdm
pyyaml
python-json-logger
uuid

# Testing
pytest
pytest-cov
pytest-mock